expand_cmd
