+++ srcs/analysis/analysis.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   analysis.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hnakayam <hnakayam@student.42.fr>          +#+  +:+       +#+        */
+/*   By: retanaka <retanaka@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/11 17:47:09 by retanaka          #+#    #+#             */
-/*   Updated: 2024/11/24 13:59:29 by hnakayam         ###   ########.fr       */
+/*   Updated: 2024/11/25 14:37:05 by retanaka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,13 +21,11 @@ t_cond	*analysis(char *src)
 	tkns = lexer(src);
 	if (!tkns)
 		return (NULL);
-	// print_tkns(tkns);
 	if (!checker(tkns))
 		return (ft_printf_stderr("minishell: syntax error\n"), NULL);
 	conds = parser(tkns);
 	free(tkns);
 	if (!conds)
 		return (ft_printf_stderr("minishell: malloc error\n"), NULL);
-	// print_conds(conds);
 	return (conds);
 }
