+++ srcs/analysis/print.c
@@ -6,7 +6,7 @@
 /*   By: retanaka <retanaka@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 14:13:57 by retanaka          #+#    #+#             */
-/*   Updated: 2024/11/14 14:23:42 by retanaka         ###   ########.fr       */
+/*   Updated: 2024/11/25 15:00:54 by retanaka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,54 +18,6 @@ void	print_word(const char *head, const char *tail)
 	if (!head)
 		return ((void)ft_printf("NULL"));
 	write(1, head, tail - head);
-}
-
-void	print_tkns(t_tkn *tkns)
-{
-	size_t	i;
-
-	ft_printf("[tkns]\n");
-	if (!tkns)
-		return ((void)ft_printf("NULL\n"));
-	i = 0;
-	ft_printf("====================\n");
-	ft_printf("=type=\t| =content=\n");
-	ft_printf("========|===========\n");
-	while (tkns[i].type != TAIL)
-	{
-		if (type_is_cmd(tkns[i].type))
-		{
-			if (tkns[i].type == NORMAL)
-				ft_printf(" normal");
-			else if (tkns[i].type == SINGLE)
-				ft_printf(" '");
-			else if (tkns[i].type == DOUBLE)
-				ft_printf(" \"");
-			else if (tkns[i].type == LESS)
-				ft_printf(" <");
-			else if (tkns[i].type == GREAT)
-				ft_printf(" >");
-			else if (tkns[i].type == DLESS)
-				ft_printf(" <<");
-			else if (tkns[i].type == DGREAT)
-				ft_printf(" >>");
-			ft_printf("\t|  ");
-			print_word(tkns[i].head, tkns[i].tail);
-			ft_printf("$\n");
-		}
-		else
-		{
-			if (tkns[i].type == AND_IF)
-				ft_printf(" &&");
-			else if (tkns[i].type == OR_IF)
-				ft_printf(" ||");
-			else if (tkns[i].type == PIPE)
-				ft_printf(" |");
-			ft_printf("\t|\n");
-		}
-		i++;
-	}
-	ft_printf("====================\n\n");
 }
 
 void	print_tkns_easy(t_tkn *tkns)
